Fix user context menu, and double-click
When you interact with a username (for example, right-clicking to bring
up a context menu, or double-click to do whatever action you have
defined in your application preferences,) Textual looks for the
"nickname" attribute of the DOM element to find the username it should
work with.

Previously, Bonfire did not set this attribute correctly. This meant
many user interactions were broken.

This commit fixes that by changing the newMessagePostedWithSender
template to add a "nickname" attribute to the username <span> that
detects double click and context menu events.